+++ src/Commands/Self/Plugin/PluginBaseCommand.php
@@ -1,0 +1,125 @@
+<?php
+
+namespace Pantheon\Terminus\Commands\Self\Plugin;
+
+use Pantheon\Terminus\Commands\TerminusCommand;
+use Pantheon\Terminus\Exceptions\TerminusNotFoundException;
+use Pantheon\Terminus\Helpers\LocalMachineHelper;
+use Pantheon\Terminus\Plugins\PluginDiscovery;
+
+/**
+ * Class PluginBaseCommand
+ * Base class for Terminus commands that deal with sending Plugin commands
+ * @package Pantheon\Terminus\Commands\Self\Plugin
+ */
+abstract class PluginBaseCommand extends TerminusCommand
+{
+    // Messages
+    const INSTALL_COMPOSER_MESSAGE =
+        'Please install Composer to enable plugin management. See https://getcomposer.org/download/.';
+    const INSTALL_GIT_MESSAGE = 'Please install Git to enable plugin management.';
+    const PROJECT_NOT_FOUND_MESSAGE = 'No project or plugin named {project} found.';
+
+    /**
+     * @var array|null
+     */
+    private $projects = null;
+
+    /**
+     * @return LocalMachineHelper
+     */
+    protected function getLocalMachine()
+    {
+        return $this->getContainer()->get(LocalMachineHelper::class);
+    }
+
+    /**
+     * Check for minimum plugin command requirements.
+     * @throws TerminusNotFoundException
+     */
+    protected function checkRequirements()
+    {
+        if (!self::commandExists('composer')) {
+            throw new TerminusNotFoundException(self::INSTALL_COMPOSER_MESSAGE);
+        }
+        if (!self::commandExists('git')) {
+            throw new TerminusNotFoundException(self::INSTALL_GIT_MESSAGE);
+        }
+    }
+
+    /**
+     * Get data on a specific installed plugin.
+     *
+     * @param string $project Name of a project or plugin
+     * @return array Plugin projects
+     * @throws TerminusNotFoundException
+     */
+    protected function getPlugin($project)
+    {
+        $matches = array_filter(
+            $this->getPluginProjects(),
+            function($plugin) use ($project) {
+                return in_array($project, [$plugin->getName(), $plugin->getPluginName(),]);
+            }
+        );
+        if (empty($matches)) {
+            throw new TerminusNotFoundException(self::PROJECT_NOT_FOUND_MESSAGE, compact('project'));
+        }
+        return array_shift($matches);
+    }
+
+    /**
+     * Get plugin projects.
+     *
+     * @return array Plugin projects
+     */
+    protected function getPluginProjects()
+    {
+        if (empty($this->projects)) {
+            $this->projects = $this->getContainer()->get(PluginDiscovery::class)->discover();
+        }
+        return $this->projects;
+    }
+
+    /**
+     * Detects whether a project/plugin is installed.
+     * @param string $project
+     * @return bool
+     */
+    protected function isInstalled($project)
+    {
+        try {
+            $this->getPlugin($project);
+        } catch (TerminusNotFoundException $e) {
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * @param string $command
+     * @return array
+     */
+    protected function runCommand($command)
+    {
+        $this->log()->debug('Running {command}...', compact('command'));
+        $results = $this->getLocalMachine()->exec($command);
+        $this->log()->debug("Returned:\n{output}", $results);
+        return $results;
+    }
+
+    /**
+     * Platform independent check whether a command exists.
+     *
+     * @param string $command Command to check
+     * @return bool True if exists, false otherwise
+     */
+    private static function commandExists($command)
+    {
+        $windows = (php_uname('s') == 'Windows NT');
+        $test_command = $windows ? 'where' : 'command -v';
+        $file = popen("$test_command $command", 'r');
+        $result = fgets($file, 255);
+        return $windows ? !preg_match('#Could not find files#', $result) : !empty($result);
+    }
+}