+++ src/Plugins/PluginInfo.php
@@ -3,15 +3,45 @@
 namespace Pantheon\Terminus\Plugins;
 
 use Consolidation\AnnotatedCommand\CommandFileDiscovery;
+use League\Container\ContainerAwareInterface;
+use League\Container\ContainerAwareTrait;
+use Pantheon\Terminus\Config\ConfigAwareTrait;
 use Pantheon\Terminus\Exceptions\TerminusException;
+use Pantheon\Terminus\Exceptions\TerminusNotFoundException;
+use Pantheon\Terminus\Helpers\LocalMachineHelper;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerAwareTrait;
+use Robo\Contract\ConfigAwareInterface;
 
 /**
  * Class PluginInfo
  * @package Pantheon\Terminus\Plugins
  */
-class PluginInfo
+class PluginInfo implements ConfigAwareInterface, ContainerAwareInterface, LoggerAwareInterface
 {
+    use ConfigAwareTrait;
+    use ContainerAwareTrait;
+    use LoggerAwareTrait;
+
     const MAX_COMMAND_DEPTH = 4;
+
+    // Commands
+    const CHANGE_DIRECTORY_AND = '[ -d {dir} ] && cd {dir} && ';
+    const GET_BRANCH_INSTALLED_VERSION_COMMAND = 'git rev-parse --abbrev-ref HEAD';
+    const GET_NONSTABLE_LATEST_VERSION_COMMAND =
+        'git tag -l --sort=version:refname | grep {version} | sort -r | xargs';
+    const GET_STABLE_LATEST_VERSION_COMMAND =
+        'git fetch --all && git tag -l --sort=version:refname | grep ^[v{version}] | sort -r | head -1';
+    const GET_TAGS_INSTALLED_VERSION_COMMAND = 'git describe --tags 2> /dev/null';
+    const VALIDATION_COMMAND = 'composer search -N -t terminus-plugin {project}';
+
+    // Installation Methods
+    const COMPOSER_METHOD = 'composer';
+    const GIT_METHOD = 'git';
+    const UNKNOWN_METHOD = 'unknown';
+
+    // Version Numbers
+    const UNKNOWN_VERSION = 'unknown';
 
     /**
      * @var null|array
@@ -21,6 +51,10 @@ class PluginInfo
      * @var string
      */
     protected $plugin_dir;
+    /**
+     * @var string
+     */
+    protected $stable_latest_version;
 
     /**
      * PluginInfo constructor.
@@ -45,7 +79,7 @@ class PluginInfo
         if ($this->usesAutoload()) {
             $info = $this->getInfo();
             foreach ($info['autoload']['psr-4'] as $prefix => $path) {
-                $loader->addPsr4($prefix, $this->plugin_dir . DIRECTORY_SEPARATOR . $path);
+                $loader->addPsr4($prefix, $this->getPath() . DIRECTORY_SEPARATOR . $path);
             }
         }
     }
@@ -98,13 +132,199 @@ class PluginInfo
         return $this->info;
     }
 
+    /**
+     * Get the currently installed plugin version.
+     *
+     * @return string Installed plugin version
+     */
+    public function getInstalledVersion()
+    {
+        try {
+            return $this->getTagInstalledVersion();
+        } catch (TerminusNotFoundException $e) {
+            try {
+                return $this->getBranchInstalledVersion();
+            } catch (TerminusNotFoundException $e) {
+                return self::UNKNOWN_VERSION;
+            }
+        }
+    }
+
+    /**
+     * Get the plugin installation method.
+     *
+     * @return string Plugin installation method
+     */
+    public function getInstallationMethod()
+    {
+        $git_dir = $this->getPath() . DIRECTORY_SEPARATOR . '.git';
+        if (is_dir($git_dir)) {
+            return self::GIT_METHOD;
+        }
+        $composer_json = $this->getPath() . DIRECTORY_SEPARATOR . 'composer.json';
+        if (file_exists($composer_json)) {
+            return self::COMPOSER_METHOD;
+        }
+        return self::UNKNOWN_METHOD;
+    }
+
+    /**
+     * Get the latest available plugin version.
+     *
+     * @return string Latest plugin version
+     */
+    public function getLatestVersion()
+    {
+        try {
+            return $this->getNonstableLatestVersion();
+        } catch (TerminusNotFoundException $e) {
+            try {
+                return $this->getBranchInstalledVersion();
+            } catch (TerminusNotFoundException $e) {
+                return self::UNKNOWN_VERSION;
+            }
+        }
+    }
+
+    /**
+     * @return string
+     */
     public function getName()
     {
         $info = $this->getInfo();
         if (isset($info['name'])) {
             return $info['name'];
         }
-        return basename($this->plugin_dir);
+        return basename($this->getPath());
+    }
+
+    /**
+     * Checks for non-stable semantic version (ie. -beta1 or -rc2).
+     * @return string The version number
+     * @throws TerminusNotFoundException If the lookup of stable version fails
+     */
+    public function getNonstableLatestVersion()
+    {
+        if (empty($this->nonstable_latest_version)) {
+            $command = str_replace(
+                ['{dir}', '{version}',],
+                [$this->getPath(), self::getMajorVersionFromVersion($this->getConfig()->get('version'))],
+                self::CHANGE_DIRECTORY_AND . self::GET_NONSTABLE_LATEST_VERSION_COMMAND,
+            );
+            $results = $this->runCommand($command);
+            $releases = self::filterForVersionNumbers($results);
+            $stable_version = $this->getStableLatestVersion();
+
+            if (count($releases) > 0) {
+                foreach ($releases as $release) {
+                    // Update to stable release, if available.
+                    if ($release === $stable_version) {
+                        return $release;
+                    }
+                }
+            }
+        }
+
+        return $this->nonstable_latest_version;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPackagistURL()
+    {
+        return 'https://packagist.org/packages/'. $this->getName();
+    }
+
+    /**
+     * @return string Location of the plugin installation
+     */
+    public function getPath()
+    {
+        return $this->plugin_dir;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPluginName()
+    {
+        return self::getPluginNameFromProjectName($this->getName());
+    }
+
+    /**
+     * @return string The version number
+     * @throws TerminusNotFoundException If the lookup fails
+     */
+    public function getStableLatestVersion()
+    {
+        if (empty($this->stable_latest_version)) {
+            $command = str_replace(
+                ['{dir}', '{version}',],
+                [$this->getPath(), self::getMajorVersionFromVersion($this->getConfig()->get('version'))],
+                self::CHANGE_DIRECTORY_AND . self::GET_STABLE_LATEST_VERSION_COMMAND,
+            );
+            $results = $this->runCommand($command);
+            $version = self::filterForVersionNumbers($results['output']);
+            if (empty($version)) {
+                throw new TerminusNotFoundException('Stable latest version not found.');
+            }
+            $this->stable_latest_version = array_shift($version);
+        }
+        return $this->stable_latest_version;
+    }
+
+    /**
+     * Check whether a Packagist project is valid.
+     *
+     * @return bool True if valid, false otherwise
+     */
+    public function isValidPackagistProject()
+    {
+        return self::checkWhetherPackagistProject($this->getName(), $this->getLocalMachine());
+    }
+
+
+    /**
+     * Check whether a Packagist project is valid.
+     *
+     * @param string $project_name Name of plugin package to install
+     * @param LocalMachineHelper $local_machine_helper
+     * @return bool True if valid, false otherwise
+     */
+    public static function checkWhetherPackagistProject($project_name, LocalMachineHelper $local_machine_helper)
+    {
+        // Search for the Packagist project.
+        $command = str_replace(
+            '{project}',
+            $project_name,
+            self::VALIDATION_COMMAND
+        );
+        $results = $local_machine_helper->exec($command);
+        $result = (trim($results['output']));
+        if (empty($result)) {
+            return false;
+        }
+        return ($result === $project_name);
+    }
+
+    /**
+     * @param $version_number
+     * @return string
+     */
+    public static function getMajorVersionFromVersion($version_number)
+    {
+        preg_match('/(\d*).\d*.\d*/', $version_number, $version_matches);
+        return $version_matches[1];
+    }
+
+    /**
+     * @return string
+     */
+    public static function getPluginNameFromProjectName($project_name)
+    {
+        preg_match('/.*\/(.*)/', $project_name, $matches);
+        return $matches[1];
     }
 
     /**
@@ -134,6 +354,14 @@ class PluginInfo
     {
         $autoload = $this->getAutoloadInfo();
         return $autoload['prefix'];
+    }
+
+    /**
+     * @return LocalMachineHelper
+     */
+    protected function getLocalMachine()
+    {
+        return $this->getContainer()->get(LocalMachineHelper::class);
     }
 
     /**
@@ -225,6 +453,26 @@ class PluginInfo
     }
 
     /**
+     * @return string The version number
+     * @throws TerminusNotFoundException If the lookup fails
+     */
+    private function getBranchInstalledVersion()
+    {
+        $command = str_replace(
+            '{dir}',
+            $this->getPath(),
+            self::CHANGE_DIRECTORY_AND . self::GET_BRANCH_INSTALLED_VERSION_COMMAND
+        );
+        $results = $this->runCommand($command);
+
+        $version = $results['output'];
+        if (empty($version)) {
+            throw new TerminusNotFoundException('Installed branch version not found.');
+        }
+        return $version;
+    }
+
+    /**
      * Return the directory where this plugin stores it's command files.
      *
      * @return string
@@ -232,6 +480,64 @@ class PluginInfo
     private function getCommandFileDirectory()
     {
         $autoload = $this->getAutoloadInfo();
-        return $this->plugin_dir . '/' . $autoload['dir'];
+        return $this->getPath() . '/' . $autoload['dir'];
+    }
+
+    /**
+     * @return string The version number
+     * @throws TerminusNotFoundException If the lookup fails
+     */
+    private function getTagInstalledVersion()
+    {
+        $command = str_replace(
+            '{dir}',
+            $this->getPath(),
+            self::CHANGE_DIRECTORY_AND . self::GET_TAGS_INSTALLED_VERSION_COMMAND
+        );
+        $results = $this->runCommand($command);
+        $version = $results['output'];
+        if (empty($version)) {
+            throw new TerminusNotFoundException('Installed tag version not found.');
+        }
+        return $version;
+    }
+
+    /**
+     * @param string $command
+     * @return array
+     */
+    private function runCommand($command)
+    {
+        $this->logger->debug('Running {command}...', compact('command'));
+        $results = $this->getLocalMachine()->exec($command);
+        $this->logger->debug("Returned:\n{output}", $results);
+        return $results;
+    }
+
+    /**
+     * @param string $results Version results from Composer
+     * @return array
+     */
+    private static function filterForVersionNumbers($results)
+    {
+        if (is_string($results)) {
+            $results = explode(PHP_EOL, $results);
+        }
+        if (empty($results)) {
+            return [];
+        }
+        return array_map(
+            function ($result) {
+                preg_match('/(\d*\.\d*\.\d*)/', $result, $output_array);
+                return $output_array[1];
+            },
+            array_filter(
+                $results,
+                function ($result) {
+                    preg_match('/(\d*\.\d*\.\d*)/', $result, $output_array);
+                    return isset($output_array[1]);
+                }
+            )
+        );
     }
 }